+++ src/main/java/io/vlingo/xoom/designer/Configuration.java
@@ -22,7 +22,7 @@ import io.vlingo.xoom.designer.task.projectgeneration.code.java.resource.RestRes
 import io.vlingo.xoom.designer.task.projectgeneration.code.java.schemata.SchemataGenerationStep;
 import io.vlingo.xoom.designer.task.projectgeneration.code.java.storage.StorageGenerationStep;
 import io.vlingo.xoom.designer.task.projectgeneration.code.java.turbosettings.TurboSettingsGenerationStep;
-import io.vlingo.xoom.designer.task.projectgeneration.code.java.unittest.entitty.EntityUnitTestGenerationStep;
+import io.vlingo.xoom.designer.task.projectgeneration.code.java.unittest.entity.EntityUnitTestGenerationStep;
 import io.vlingo.xoom.designer.task.projectgeneration.code.java.unittest.queries.QueriesUnitTestGenerationStep;
 import io.vlingo.xoom.designer.task.projectgeneration.code.reactjs.AggregateGenerationStep;
 import io.vlingo.xoom.designer.task.projectgeneration.code.reactjs.LayoutGenerationStep;
