+++ src/main/java/io/vlingo/xoom/designer/task/projectgeneration/code/java/unittest/entity/MockDispatcherDetail.java
@@ -5,7 +5,7 @@
 // was not distributed with this file, You can obtain
 // one at https://mozilla.org/MPL/2.0/.
 
-package io.vlingo.xoom.designer.task.projectgeneration.code.java.unittest.entitty;
+package io.vlingo.xoom.designer.task.projectgeneration.code.java.unittest.entity;
 
 public class MockDispatcherDetail {
