+++ src/main/java/io/vlingo/xoom/designer/task/projectgeneration/code/java/unittest/entity/ResultAssignmentStatement.java
@@ -5,12 +5,12 @@
 // was not distributed with this file, You can obtain
 // one at https://mozilla.org/MPL/2.0/.
 
-package io.vlingo.xoom.designer.task.projectgeneration.code.java.unittest.entitty;
+package io.vlingo.xoom.designer.task.projectgeneration.code.java.unittest.entity;
 
 import io.vlingo.xoom.codegen.content.CodeElementFormatter;
 import io.vlingo.xoom.codegen.parameter.CodeGenerationParameter;
-import io.vlingo.xoom.designer.task.projectgeneration.code.java.JavaTemplateStandard;
 import io.vlingo.xoom.designer.task.projectgeneration.Label;
+import io.vlingo.xoom.designer.task.projectgeneration.code.java.JavaTemplateStandard;
 
 import java.util.stream.Collectors;
